Prevent the local bindings of `make-eval-tree' from shadowing the form
Emacs' dynamic scoping can be troublesome when evaluating a form
defined outside the current scope. If any of `make-eval-tree's local
variables occur in the FORM argument, they could influence the
evaluation of the form.

The solution used here is the same as for macros: make an uninterned
symbol with `make-symbol' and substitute that into the function
definition. The resulting (eval `(defun ...)) code isn't very pretty;
if there is a cleaner solution, I'd like to know.